# Caching performance improvements