### Description

Caching performance improvements and benchmarks

* 50% improvement for `cache.AddType` on existing types
* 10% improvement for `cache.AddType` for new types
* 10% improvement for `cache.AddIndex`
* 50% reduction of size of `kindIndices`

old:
```
go test -bench Benchmark_cache_AddIndex -run ^$ -count 5 ./internal/controller/cache
goos: darwin
goarch: arm64
pkg: github.com/hashicorp/consul/internal/controller/cache
Benchmark_cache_AddIndex-12      1600906               732.0 ns/op
Benchmark_cache_AddIndex-12      1622610               778.1 ns/op
Benchmark_cache_AddIndex-12      1922361               856.4 ns/op
Benchmark_cache_AddIndex-12      1581726               737.5 ns/op
Benchmark_cache_AddIndex-12      1658888               809.6 ns/op
PASS
ok      github.com/hashicorp/consul/internal/controller/cache   29.278s

go test -bench Benchmark_cache_AddType -run ^$ -count 5 -benchmem ./internal/controller/cache
goos: darwin
goarch: arm64
pkg: github.com/hashicorp/consul/internal/controller/cache
Benchmark_cache_AddType_New-12           1757521               773.1 ns/op           790 B/op          9 allocs/op
Benchmark_cache_AddType_New-12           1915255               743.8 ns/op           772 B/op          9 allocs/op
Benchmark_cache_AddType_New-12           1684995               844.1 ns/op           800 B/op          9 allocs/op
Benchmark_cache_AddType_New-12           1840144               741.4 ns/op           780 B/op          9 allocs/op
Benchmark_cache_AddType_New-12           1781332               789.1 ns/op           787 B/op          9 allocs/op
Benchmark_cache_AddType_Existing-12     34208458                35.75 ns/op            0 B/op          0 allocs/op
Benchmark_cache_AddType_Existing-12     34524012                35.45 ns/op            0 B/op          0 allocs/op
Benchmark_cache_AddType_Existing-12     34367828                35.03 ns/op            0 B/op          0 allocs/op
Benchmark_cache_AddType_Existing-12     34876370                36.42 ns/op            0 B/op          0 allocs/op
Benchmark_cache_AddType_Existing-12     34563870                34.63 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/hashicorp/consul/internal/controller/cache   31.074s
```

new:
```
go test -bench Benchmark_cache_AddIndex -run ^$ -count 5 ./internal/controller/cache
goos: darwin
goarch: arm64
pkg: github.com/hashicorp/consul/internal/controller/cache
Benchmark_cache_AddIndex-12      2092167               702.0 ns/op
Benchmark_cache_AddIndex-12      1912968               708.4 ns/op
Benchmark_cache_AddIndex-12      1838739               738.5 ns/op
Benchmark_cache_AddIndex-12      1727347               757.3 ns/op
Benchmark_cache_AddIndex-12      1931530               703.1 ns/op
PASS
ok      github.com/hashicorp/consul/internal/controller/cache   27.553s

go test -bench Benchmark_cache_AddType -run ^$ -count 5 -benchmem ./internal/controller/cache
goos: darwin
goarch: arm64
pkg: github.com/hashicorp/consul/internal/controller/cache
Benchmark_cache_AddType_New-12           1810838               685.5 ns/op           783 B/op          9 allocs/op
Benchmark_cache_AddType_New-12           1774274               759.1 ns/op           788 B/op          9 allocs/op
Benchmark_cache_AddType_New-12           2014064               692.9 ns/op           762 B/op          9 allocs/op
Benchmark_cache_AddType_New-12           1886276               708.6 ns/op           775 B/op          9 allocs/op
Benchmark_cache_AddType_New-12           1704679               816.6 ns/op           797 B/op          9 allocs/op
Benchmark_cache_AddType_Existing-12     65070776                18.18 ns/op            0 B/op          0 allocs/op
Benchmark_cache_AddType_Existing-12     65581833                18.23 ns/op            0 B/op          0 allocs/op
Benchmark_cache_AddType_Existing-12     65527668                18.07 ns/op            0 B/op          0 allocs/op
Benchmark_cache_AddType_Existing-12     67483812                18.10 ns/op            0 B/op          0 allocs/op
Benchmark_cache_AddType_Existing-12     65270455                18.20 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/hashicorp/consul/internal/controller/cache   30.591s
```

### PR Checklist

* [ ] updated test coverage
* [ ] external facing docs updated
* [ ] appropriate backport labels added
* [ ] not a security concern